crew/requirements.txt
# Core dependencies
crewai==0.80.0
crewai-tools==0.14.0

# LangChain ecosystem
langchain==0.3.7
langchain-core==0.3.19
langchain-community==0.3.7
langchain-openai==0.2.9
langchain-experimental==0.3.3
langchain-text-splitters==0.3.2
langchain-cohere==0.3.1

# OpenAI
openai==1.54.5

# Database and API
supabase==1.2.0
postgrest==0.11.0  # Ajustado para ser compatível com supabase 1.2.0
gotrue==1.1.1
httpx>=0.24.1,<1.0.0  # Especificando limite superior

# Utils
python-dotenv==1.0.1
pandas==2.1.4
